backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
groq>=0.9.0
pinecone>=5.0.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0
google-generativeai>=0.7.0
